aider: Added a stop button to the client index page.
# Aider chat conversation:

USER: Can you add a stop button to the page

ASSISTANT: To add a stop button to the page, we would need to modify the HTML file where the page layout is defined. However, the file `client/index.html` is currently read-only.

Could you please add `client/index.html` to the chat so that we can make changes to it?

USER: I added these *read-write* files: client/index.html

ASSISTANT: To add a stop button to the page, we need to modify `client/index.html`:

1. Add a new button element with the label "Stop" after the "Find Path" button.

Here is the *SEARCH/REPLACE* block:

client/index.html
```html
<<<<<<< SEARCH
        <input type="submit" value="Find Path">
=======
        <input type="submit" value="Find Path">
        <button id="stop-button">Stop</button>
>>>>>>> REPLACE
```