aider: Replaced hardcoded keywords with words extracted from start and finish page URLs in link ranking heuristic.
# Aider chat conversation:

USER: the keywords should be defined based on the start and finish wikipedia pages

ASSISTANT: To define the keywords based on the start and finish Wikipedia pages, we can extract the last part of the URL (after the last '/') and split it into words. These words can be used as the keywords for the link ranking heuristic.

Here is the *SEARCH/REPLACE* block:

server/crawler.py
```python
<<<<<<< SEARCH
        # Define your keywords
        keywords = ['keyword1', 'keyword2', 'keyword3']
=======
        # Define your keywords based on the start and finish pages
        start_keywords = start_page.split('/')[-1].split('_')
        finish_keywords = finish_page.split('/')[-1].split('_')
        keywords = start_keywords + finish_keywords
>>>>>>> REPLACE
```

This change will replace the hardcoded keywords with the words extracted from the start and finish page URLs. The '_' character is used as the delimiter because in Wikipedia URLs, spaces in the page titles are replaced with underscores.