aider: Fixed unpacking error in find_path function by adjusting the number of variables to match the returned values from get_links function.
# Aider chat conversation:

USER: what is going on with this Error occurred: too many values to unpack (expected 3)

ASSISTANT: The error message "too many values to unpack (expected 3)" indicates that a function is returning more values than you are providing variables to store them. In this case, the `get_links` function is returning three values, but when it is called in the `find_path` function, four variables are provided to store the returned values.

To fix this, we need to modify the `find_path` function in the `server/crawler.py` file to correctly unpack the values returned by the `get_links` function.

Here is the *SEARCH/REPLACE* block:

server/crawler.py
```python
<<<<<<< SEARCH
            start_links, start_text, start_categories = get_links(start_page, start_page, finish_page, category_dict)
            finish_links, finish_text, finish_categories = get_links(finish_page, start_page, finish_page, category_dict)
=======
            start_links, start_text = get_links(start_page, start_page, finish_page, category_dict)
            finish_links, finish_text = get_links(finish_page, start_page, finish_page, category_dict)
>>>>>>> REPLACE
```