aider: Updated find_path function to include intermediate pages in the path from start to finish.
# Aider chat conversation:

USER: if two of the wikipedia paths have a common page, then does it mean that the start and finish pages are connected?

ASSISTANT: Yes, if two paths from the start page and the finish page have a common page, it means that the start and finish pages are connected. In the context of the Wikipedia pages, a connection between two pages is established if there is a sequence of links that you can follow from the start page to the finish page. If a page is common between the paths from the start and finish pages, you can navigate from the start page to this common page, and then from the common page to the finish page. Therefore, the start and finish pages are connected.

USER: but does it mean that you can navigate from the start page to the finish page?

ASSISTANT: Yes, if two paths from the start page and the finish page have a common page, it means that you can navigate from the start page to the finish page.

In the context of the Wikipedia pages, a connection between two pages is established if there is a sequence of links that you can follow from the start page to the finish page. If a page is common between the paths from the start and finish pages, you can navigate from the start page to this common page, and then from the common page to the finish page. Therefore, the start and finish pages are connected.

However, it's important to note that the path found may not be the shortest path between the start and finish pages. The algorithm used in the `find_path` function is a bidirectional search, which doesn't necessarily find the shortest path. It stops as soon as it finds any path between the start and finish pages. If you need the shortest path, you would need to use a different algorithm, such as Dijkstra's algorithm or A*.

USER: it still only says the start and finish page in the path

ASSISTANT: To fix the issue where the path only includes the start and finish page, we need to modify the `find_path` function in `server/crawler.py` to correctly construct the path from start to finish.

Currently, the path is constructed by concatenating `path_start` and `path_finish[::-1]`. However, this only includes the start and finish pages. To include the intermediate pages in the path, we need to update `path_start` and `path_finish` with the next page in the path when we add a page to the queue.

Here are the *SEARCH/REPLACE* blocks:

server/crawler.py
```python
<<<<<<< SEARCH
queue_start.append((next_start, path_start + [next_start], depth_start + 1))
=======
queue_start.append((next_start, path_start + [next_start], depth_start + 1))
path_start.append(next_start)
>>>>>>> REPLACE
```

server/crawler.py
```python
<<<<<<< SEARCH
queue_finish.append((next_finish, path_finish + [next_finish], depth_finish + 1))
=======
queue_finish.append((next_finish, path_finish + [next_finish], depth_finish + 1))
path_finish.append(next_finish)
>>>>>>> REPLACE
```

Please note that this will include the intermediate pages in the path in the order they are discovered, not necessarily in the order they are visited. If you want the path to reflect the order of visitation, you would need to modify the algorithm to track the order of visitation.