aider: Modified find_path function to append only connecting links to path_start and path_finish.
# Aider chat conversation:

USER: Is it still using keywords from the title?

ASSISTANT: Yes, the current implementation in `server/crawler.py` is still using keywords from the title of the start and finish pages.

This is evident in the `find_path` function where `start_keywords` and `finish_keywords` are derived from the last part of the start and finish page URLs respectively. These keywords are then combined to form the `keywords` list which is used in the `get_links` function to score the links.

Here is the relevant code:

```python
# Define your keywords based on the start and finish pages
start_keywords = start_page.split('/')[-1].split('_')
finish_keywords = finish_page.split('/')[-1].split('_')
keywords = start_keywords + finish_keywords
```

This code splits the start and finish page URLs on the '/' character and takes the last part (which is the title of the page). It then splits this title on the '_' character to get the individual words in the title. These words are used as the keywords for scoring the links.

USER: why does it seem like a DFS? I am getting extremely long paths from start page to end page

ASSISTANT: The current implementation in `server/crawler.py` is using a bidirectional search with priority queues (implemented with heaps) for both the start and finish sides. However, the way the priority (score) is calculated and used might be causing the search to behave more like a Depth-First Search (DFS) rather than a Breadth-First Search (BFS).

In the current implementation, the score for a link is calculated based on the number of keywords it contains and whether it is in the category of the finish page. The links are then added to the priority queue with their scores negated. This means that links with higher scores (i.e., those containing more keywords or in the category of the finish page) are given higher priority and are explored first. This could lead to the search going deeper into the graph (like DFS) if such links lead to paths that are not necessarily the shortest.

To make the search behave more like a BFS and thus find shorter paths, you could modify the scoring system to give higher priority to links that are closer to the start page. One way to do this could be to incorporate the depth of the link (i.e., the number of links that have to be followed from the start page to reach this link) into the score. For example, you could subtract the depth from the current score when adding a link to the priority queue. This would give higher priority to links that are closer to the start page, making the search explore these links first and thus behave more like a BFS.

Please let me know if you want to proceed with this change and I will provide the necessary code modifications.

USER: I want the path from start page to finish page, not all the pages visited

ASSISTANT: To modify the code to return only the path from the start page to the finish page, we need to modify the `find_path` function in `server/crawler.py`.

Currently, the `find_path` function is appending every visited link to `path_start` and `path_finish`. Instead, we should only append a link to the path when it is also in the other direction's discovered set. This means it is a connecting link between the two search directions and should be part of the final path.

Here are the *SEARCH/REPLACE* blocks:

server/crawler.py
```python
<<<<<<< SEARCH
                        path_start.append(next_start)
=======
                        if next_start in discovered_finish:
                            path_start.append(next_start)
>>>>>>> REPLACE
```

server/crawler.py
```python
<<<<<<< SEARCH
                        path_finish.append(next_finish)
=======
                        if next_finish in discovered_start:
                            path_finish.append(next_finish)
>>>>>>> REPLACE
```

These changes will ensure that only the links that form the path from the start page to the finish page are appended to `path_start` and `path_finish`. The final path is then formed by concatenating `path_start` and the reverse of `path_finish`, which should now give you the desired path from the start page to the finish page.